+++ tools/src/main/java/it/eng/tools/rest/api/ApplicationPropertiesAPIController.java
@@ -3,7 +3,6 @@ package it.eng.tools.rest.api;
 import java.util.List;
 import java.util.Optional;
 
-import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -14,12 +13,10 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.fasterxml.jackson.databind.JsonNode;
-
 import it.eng.tools.exception.ApplicationPropertyNotChangedAPIException;
 import it.eng.tools.exception.ApplicationPropertyNotFoundAPIException;
 import it.eng.tools.model.ApplicationProperty;
-import it.eng.tools.model.Serializer;
+import it.eng.tools.response.GenericApiResponse;
 import it.eng.tools.service.ApplicationPropertiesService;
 import lombok.extern.java.Log;
 
@@ -31,8 +28,6 @@ import lombok.extern.java.Log;
 @Log
 public class ApplicationPropertiesAPIController {
 
-	private final ApplicationEventPublisher applicationEventPublisher;
-
 	private final ApplicationPropertiesService propertiesService;
 
 	/**
@@ -40,10 +35,9 @@ public class ApplicationPropertiesAPIController {
 	 * @param service ApplicationPropertiesService
 	 * @param applicationEventPublisher ApplicationEventPublisher
 	 */
-	public ApplicationPropertiesAPIController(ApplicationPropertiesService service, ApplicationEventPublisher applicationEventPublisher) {
+	public ApplicationPropertiesAPIController(ApplicationPropertiesService service) {
 		super();
 		this.propertiesService = service;
-		this.applicationEventPublisher = applicationEventPublisher;
 	}
 
 	/**
@@ -52,15 +46,14 @@ public class ApplicationPropertiesAPIController {
 	 * @return List of properties
 	 */
 	@GetMapping(path = "/")
-	public ResponseEntity<List<JsonNode>> getProperties(@RequestParam(required = false) String key_prefix) {
+	public ResponseEntity<GenericApiResponse<List<ApplicationProperty>>> getProperties(@RequestParam(required = false) String key_prefix) {
 		log.info("getProperties()");
 		if(key_prefix != null && !key_prefix.isBlank()) log.info(" with key_prefix " + key_prefix);
 		var properties = propertiesService.getProperties(key_prefix);
 
-		applicationEventPublisher.publishEvent(properties);
-
+		GenericApiResponse<List<ApplicationProperty>> genericApiResponse = GenericApiResponse.success(properties, "Application properties with prefix");
 		return ResponseEntity.ok().contentType(MediaType.APPLICATION_JSON)
-				.body(Serializer.serializeProtocolListOfJsonNode(properties));
+				.body(genericApiResponse);
 	}
 
 	/**
@@ -69,14 +62,13 @@ public class ApplicationPropertiesAPIController {
 	 * @return property
 	 */
 	@GetMapping(path = "/{key}")
-	public ResponseEntity<JsonNode> getPropertyByKey(@PathVariable String key) {
+	public ResponseEntity<GenericApiResponse<ApplicationProperty>> getPropertyByKey(@PathVariable String key) {
 		log.info("Fetching property with key " + key);
 
 		ApplicationProperty property = propertiesService.getPropertyByKey(key).orElseThrow(() -> new ApplicationPropertyNotFoundAPIException("Property with key " + key + " not found"));
 
-		applicationEventPublisher.publishEvent(property);
 		return ResponseEntity.ok().contentType(MediaType.APPLICATION_JSON)
-				.body(Serializer.serializeProtocolJsonNode(property));
+				.body(GenericApiResponse.success(property, "Application property for key"));
 	}
 
 	/**
@@ -85,7 +77,7 @@ public class ApplicationPropertiesAPIController {
 	 * @return Response
 	 */
 	@PutMapping(path = "/")
-	public ResponseEntity<JsonNode> modifyProperty(@RequestBody ApplicationProperty property) {
+	public ResponseEntity<GenericApiResponse<ApplicationProperty>> modifyProperty(@RequestBody ApplicationProperty property) {
 		log.info("modifyProperty(...) ");
 		log.info("property = " + property);
 
@@ -102,7 +94,7 @@ public class ApplicationPropertiesAPIController {
 		}
 
 		return ResponseEntity.ok().contentType(MediaType.APPLICATION_JSON)
-				.body(Serializer.serializeProtocolJsonNode(storedProperty));
+				.body(GenericApiResponse.success(storedProperty, "Application property updated"));
 	}
 
 }